+++ api-base/types/index.ts
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // These are augmented, do an augmentation export
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/calls.ts';
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/consts.ts';
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/errors.ts';
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/events.ts';
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/storage.ts';
-export * from 'https://deno.land/x/polkadot@0.2.25/api-base/types/submittable.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/calls.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/consts.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/errors.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/events.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/storage.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/submittable.ts';
 
 // normal exports
 export * from './api.ts';