+++ api-derive/bundle.ts
@@ -1,8 +1,8 @@
 // Copyright 2017-2023 @polkadot/api-derive authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { DeriveCustom } from 'https://deno.land/x/polkadot@0.2.25/api-base/types/index.ts';
-import type { AnyFunction, AnyString } from 'https://deno.land/x/polkadot@0.2.25/types/types/index.ts';
+import type { DeriveCustom } from 'https://deno.land/x/polkadot/api-base/types/index.ts';
+import type { AnyFunction, AnyString } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ExactDerive } from './derive.ts';
 import type { DeriveApi } from './types.ts';