+++ types/codec/types.ts
@@ -1,4 +1,4 @@
 // Copyright 2017-2023 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export type { U8aBitLength, UIntBitLength } from 'https://deno.land/x/polkadot@0.2.25/types-codec/types/index.ts';
+export type { U8aBitLength, UIntBitLength } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';